tech(e2e): extract user data used for fixtures in separate file
