+++ high-level-tests/e2e-playwright/helpers/certification/data.ts
@@ -1,0 +1,67 @@
+export const pixCertifiableUserData = [
+  {
+    firstName: 'Buffy',
+    lastName: 'Summers',
+    rawPassword: 'pix123',
+    sex: 'F',
+    birthdate: '1981-01-19',
+    birthDay: '19',
+    birthMonth: '01',
+    birthYear: '1981',
+    birthCountry: 'FRANCE',
+    birthCity: 'Perpignan',
+    postalCode: '66000',
+  },
+  {
+    firstName: 'Rupert',
+    lastName: 'Giles',
+    rawPassword: 'pix123',
+    sex: 'M',
+    birthdate: '1955-02-22',
+    birthDay: '22',
+    birthMonth: '02',
+    birthYear: '1955',
+    birthCountry: 'FRANCE',
+    birthCity: 'Perpignan',
+    postalCode: '66000',
+  },
+  {
+    firstName: 'Cordelia',
+    lastName: 'Chase',
+    rawPassword: 'pix123',
+    sex: 'F',
+    birthdate: '1981-08-05',
+    birthDay: '05',
+    birthMonth: '08',
+    birthYear: '1981',
+    birthCountry: 'FRANCE',
+    birthCity: 'Perpignan',
+    postalCode: '66000',
+  },
+  {
+    firstName: 'Willow',
+    lastName: 'Rosenberg',
+    rawPassword: 'pix123',
+    sex: 'F',
+    birthdate: '1983-12-15',
+    birthDay: '15',
+    birthMonth: '12',
+    birthYear: '1983',
+    birthCountry: 'FRANCE',
+    birthCity: 'Perpignan',
+    postalCode: '66000',
+  },
+  {
+    firstName: 'Riley',
+    lastName: 'Finn',
+    rawPassword: 'pix123',
+    sex: 'M',
+    birthdate: '1981-09-29',
+    birthDay: '29',
+    birthMonth: '09',
+    birthYear: '1981',
+    birthCountry: 'FRANCE',
+    birthCity: 'Perpignan',
+    postalCode: '66000',
+  },
+];
